+++ tests/Kiota.Builder.Tests/Refiners/PhpLanguageRefinerTests.cs
@@ -235,4 +235,57 @@ public class PhpLanguageRefinerTests
         await ILanguageRefiner.Refine(new GenerationConfiguration { Language = GenerationLanguage.PHP }, root);
         Assert.Equal(2, modelClass.Usings.Count());
     }
+
+    [Fact]
+    public async Task RenamesComposedTypeWrapperWhenSimilarClassExistsInNamespace()
+    {
+        var model = new CodeClass
+        {
+            Name = "Union",
+            Kind = CodeClassKind.Model
+        };
+        var parent = new CodeClass
+        {
+            Name = "Parent"
+        };
+        root.AddClass(model, parent);
+
+        var composedType = new CodeUnionType { Name = "Union" };
+        composedType.AddType(new CodeType { Name = "string" }, new CodeType { Name = "int" });
+
+        parent.AddProperty(new CodeProperty
+        {
+            Name = "property",
+            Type = composedType
+        });
+        await ILanguageRefiner.Refine(new GenerationConfiguration { Language = GenerationLanguage.PHP }, root);
+        Assert.NotNull(root.FindChildByName<CodeClass>("UnionWrapper", false));
+    }
+
+    [Fact]
+    public async Task DoesNotCreateDuplicateComposedTypeWrapperIfOneAlreadyExists()
+    {
+        var composedType = new CodeUnionType { Name = "Union" };
+        composedType.AddType(new CodeType { Name = "string" }, new CodeType { Name = "int" });
+
+        var parent = new CodeClass
+        {
+            Name = "Parent"
+        };
+        parent.AddProperty(new CodeProperty
+        {
+            Name = "property",
+            Type = composedType
+        });
+        parent.AddProperty(new CodeProperty
+        {
+            Name = "property2",
+            Type = composedType
+        });
+        root.AddClass(parent);
+
+        await ILanguageRefiner.Refine(new GenerationConfiguration { Language = GenerationLanguage.PHP }, root);
+        Assert.True(root.FindChildByName<CodeClass>("Union", false) is CodeClass unionTypeWrapper && unionTypeWrapper.OriginalComposedType != null);
+        Assert.True(root.FindChildByName<CodeClass>("UnionWrapper", false) is null);
+    }
 }